fixing dialog

+++ installation_scripts/install.sh
@@ -35,29 +35,34 @@ clear
 shopt -s nocasematch
 if [[ $i3 =~ y ]]
 then
-	pacman -S i3-wm
+	sudo pacman -S i3-wm
 	ln -sf $LINKDOT/flavours/i3 /home/$USER/.config/
         
 else
 	echo "-- moving to qtile"
 fi
 
-read -p "-- Would you like to install Qtile?" yna
-case $yna in
-            [Yy]* ) pacman -S qtile
-		    ln -sf $LINKDOT/flavours/qtile /home/$USER/.config/
-		                       ;;
-                        * ) echo "-- if you didn't install i3 or this, you're on your own for a GUI.";;
-esac
+qtile=$(dialog --stdout --inputbox "install qtile? [y/N]" 0 0) || exit 1
+clear
+shopt -s nocasematch
+if [[ $i3 =~ y ]]
+then
+	sudo pacman -S qtile
+	ln -sf $LINKDOT/flavours/qtile /home/$USER/.config/
+else 
+	echo "-- if you didn't install i3 or this, you're on your own for a GUI.";;
+fi
 
 ln -sf $LINKDOT/.config/* /home/$USER/.config/
-read -p "-- enter sudo password to cp grub theme and sddm theme to correct locations and fix the config files. otherwise skip configuring both. Understand? [y/N] " yna
-case $yna in
-	[Yy]* ) sudo cp $LINKDOT/boot/grub/grubel /boot/grub/
-		sudo cp $LINKDOT/boot/sddm/themes/pixarch_sddm /usr/share/sddm/themes/
+theme=$(dialog --stdout --inputbox "enter sudo password to cp grub theme and sddm theme to correct locations and fix the config files. otherwise skip configuring both. Understand? [y/N]" 0 0) || exit 1
+if [[ $theme =~ y ]]
+then
+	        sudo cp -r $LINKDOT/boot/grub/grubel /boot/grub/
+		sudo cp -r $LINKDOT/boot/sddm/themes/pixarch_sddm /usr/share/sddm/themes/
 		sudo sed 's/\#GRUB_THEME/GRUB_THEME=\"\/boot\/grub\/grubel\/theme.txt\"/' -i /mnt/etc/default/grub
-                sudo echo "'[Theme]
-			   pixarch_sddm' >> /etc/sddm.conf.d/theme.conf" ;;
-			* ) echo "-- you're on your own for theming.";;
-esac
+                su -c echo "[Theme]
+			   pixarch_sddm" >> /etc/sddm.conf.d/theme.conf"
+else 
+	echo "-- you're on your own for theming."
+fi
